Initial commit - working version
working version that prints a 10x10 grid and that moves a blue cursor and paints cells black, using the WASD + Spacebar keys. Needs extensive commenting and beautifying (as it is the version created at the bootcamp) so it several commented notes unnecessary for a third party to understand.
My plan with this project is to in an initial stage, add another color, then turn this into a proper map editor, where you can paint the grid with land or sea sprites.